RST-formatted man page for security-secrets-setup.
security-secrets-setup is a new Fuji utility that enhances
the pkisetup utility from Edinburgh with explicit capabilities
to create, create and cache, or import a PKI and deploy it
on-device.  It also replaces some amount of wrapper logic
previously written in shell script by providing the higher
level abstraction.

security-secrets-setup is a critical component for the
secret management hardening being done in Fuji release.

The included man-style page documents the externally-visible
functionality of security-secrets-setup.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>