fix: add version to swagger links
The default links to SwaggerHub on the documentation pages for Hanoi
don't include a specific version, so if someone clicks them, they
get redirected to the Ireland versions of the documentation.

Note - all the links have been updated to reference version 1.2.1
*except* for the App Functions SDK page, as the latest version
published on SwaggerHub for the App Functions SDK is 1.1.0.

Signed-off-by: Tony Espy <espy@canonical.com>